requirements.txt
Django
psycopg2-binary
cryptography
PyJWT
social-auth-app-django
requests
djangorestframework
flask
flask_restful
openpyxl